修正MessageDispatcherComponent拼写错误
