+++ Server/Hotfix/Module/Message/InnerMessageDispatcher.cs
@@ -61,7 +61,7 @@ namespace ETHotfix
 				}
 				default:
 				{
-					Game.Scene.GetComponent<MessageDispatherComponent>().Handle(session, new MessageInfo(opcode, message));
+					Game.Scene.GetComponent<MessageDispatcherComponent>().Handle(session, new MessageInfo(opcode, message));
 					break;
 				}
 			}
