+++ Server/Hotfix/Module/Message/OuterMessageDispatcher.cs
@@ -37,7 +37,7 @@ namespace ETHotfix
 					}
 				}
 
-				Game.Scene.GetComponent<MessageDispatherComponent>().Handle(session, new MessageInfo(opcode, message));
+				Game.Scene.GetComponent<MessageDispatcherComponent>().Handle(session, new MessageInfo(opcode, message));
 			}
 			catch (Exception e)
 			{
