+++ Server/Model/Module/Message/MessageDispatcherComponent.cs
@@ -5,7 +5,7 @@ namespace ETModel
 	/// <summary>
 	/// 消息分发组件
 	/// </summary>
-	public class MessageDispatherComponent : Component
+	public class MessageDispatcherComponent : Component
 	{
 		public readonly Dictionary<ushort, List<IMHandler>> Handlers = new Dictionary<ushort, List<IMHandler>>();
 	}
