+++ Unity/Assets/Hotfix/Init.cs
@@ -18,7 +18,7 @@ namespace ETHotfix
 				
 				Game.Scene.AddComponent<UIComponent>();
 				Game.Scene.AddComponent<OpcodeTypeComponent>();
-				Game.Scene.AddComponent<MessageDispatherComponent>();
+				Game.Scene.AddComponent<MessageDispatcherComponent>();
 
 				// 加载热更配置
 				ETModel.Game.Scene.GetComponent<ResourcesComponent>().LoadBundle("config.unity3d");
