+++ Unity/Assets/Hotfix/Module/Message/MessageDispatcherComponent.cs
@@ -1,21 +1,22 @@
 ﻿using System;
 using System.Collections.Generic;
+using ETModel;
 
-namespace ETModel
+namespace ETHotfix
 {
 	[ObjectSystem]
-	public class MessageDispatherComponentAwakeSystem : AwakeSystem<MessageDispatherComponent>
+	public class MessageDispatcherComponentAwakeSystem : AwakeSystem<MessageDispatcherComponent>
 	{
-		public override void Awake(MessageDispatherComponent t)
+		public override void Awake(MessageDispatcherComponent self)
 		{
-			t.Awake();
+			self.Awake();
 		}
 	}
 
 	[ObjectSystem]
-	public class MessageDispatherComponentLoadSystem : LoadSystem<MessageDispatherComponent>
+	public class MessageDispatcherComponentLoadSystem : LoadSystem<MessageDispatcherComponent>
 	{
-		public override void Load(MessageDispatherComponent self)
+		public override void Load(MessageDispatcherComponent self)
 		{
 			self.Load();
 		}
@@ -24,7 +25,7 @@ namespace ETModel
 	/// <summary>
 	/// 消息分发组件
 	/// </summary>
-	public class MessageDispatherComponent : Component
+	public class MessageDispatcherComponent : Component
 	{
 		private readonly Dictionary<ushort, List<IMHandler>> handlers = new Dictionary<ushort, List<IMHandler>>();
 
@@ -37,7 +38,10 @@ namespace ETModel
 		{
 			this.handlers.Clear();
 
-			List<Type> types = Game.EventSystem.GetTypes(typeof(MessageHandlerAttribute));
+			ETModel.MessageDispatcherComponent messageDispatcherComponent = ETModel.Game.Scene.GetComponent<ETModel.MessageDispatcherComponent>();
+			ETModel.OpcodeTypeComponent opcodeTypeComponent = ETModel.Game.Scene.GetComponent<ETModel.OpcodeTypeComponent>();
+
+			List<Type> types = Game.EventSystem.GetTypes();
 
 			foreach (Type type in types)
 			{
@@ -53,15 +57,26 @@ namespace ETModel
 					Log.Error($"message handle {type.Name} 需要继承 IMHandler");
 					continue;
 				}
-
+				
 				Type messageType = iMHandler.GetMessageType();
 				ushort opcode = this.Entity.GetComponent<OpcodeTypeComponent>().GetOpcode(messageType);
-				if (opcode == 0)
+				if (opcode != 0)
 				{
-					Log.Error($"消息opcode为0: {messageType.Name}");
-					continue;
+					this.RegisterHandler(opcode, iMHandler);
 				}
-				this.RegisterHandler(opcode, iMHandler);
+
+				// 尝试注册到mono层
+				if (messageDispatcherComponent != null && opcodeTypeComponent != null)
+				{
+					ushort monoOpcode = opcodeTypeComponent.GetOpcode(messageType);
+					if (monoOpcode == 0)
+					{
+						continue;
+					}
+
+					MessageProxy messageProxy = new MessageProxy(messageType, (session, o) => { iMHandler.Handle(session, o); });
+					messageDispatcherComponent.RegisterHandler(monoOpcode, messageProxy);
+				}
 			}
 		}
 
@@ -74,12 +89,12 @@ namespace ETModel
 			this.handlers[opcode].Add(handler);
 		}
 
-		public void Handle(Session session, MessageInfo messageInfo)
+		public void Handle(ETModel.Session session, MessageInfo messageInfo)
 		{
 			List<IMHandler> actions;
 			if (!this.handlers.TryGetValue(messageInfo.Opcode, out actions))
 			{
-				Log.Error($"消息没有处理: {messageInfo.Opcode} {JsonHelper.ToJson(messageInfo.Message)}");
+				Log.Error($"消息 {messageInfo.Message.GetType().FullName} 没有处理");
 				return;
 			}
