+++ Unity/Assets/Hotfix/Module/Message/MessageDispatcherComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c5e3350a88eaef468e363ce00f9bd3e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
