+++ Unity/Assets/Hotfix/Module/Message/Session.cs
@@ -79,7 +79,7 @@ namespace ETHotfix
 				return;
 			}
 
-			Game.Scene.GetComponent<MessageDispatherComponent>().Handle(session, new MessageInfo(opcode, message));
+			Game.Scene.GetComponent<MessageDispatcherComponent>().Handle(session, new MessageInfo(opcode, message));
 		}
 
 		public void Send(IMessage message)
