+++ Unity/Assets/Model/ILBinding/ETModel_Entity_Binding.cs
@@ -91,7 +91,7 @@ namespace ILRuntime.Runtime.Generated
                     }
                 }
             }
-            args = new Type[]{typeof(ETModel.MessageDispatherComponent)};
+            args = new Type[]{typeof(ETModel.MessageDispatcherComponent)};
             if (genericMethods.TryGetValue("GetComponent", out lst))
             {
                 foreach(var m in lst)
@@ -270,7 +270,7 @@ namespace ILRuntime.Runtime.Generated
             ETModel.Entity instance_of_this_method = (ETModel.Entity)typeof(ETModel.Entity).CheckCLRTypes(StackObject.ToObject(ptr_of_this_method, __domain, __mStack));
             __intp.Free(ptr_of_this_method);
 
-            var result_of_this_method = instance_of_this_method.GetComponent<ETModel.MessageDispatherComponent>();
+            var result_of_this_method = instance_of_this_method.GetComponent<ETModel.MessageDispatcherComponent>();
 
             object obj_result_of_this_method = result_of_this_method;
             if(obj_result_of_this_method is CrossBindingAdaptorType)
