+++ Unity/Assets/Model/Module/Message/MessageDispatcherComponent.cs
@@ -1,22 +1,21 @@
 ﻿using System;
 using System.Collections.Generic;
-using ETModel;
 
-namespace ETHotfix
+namespace ETModel
 {
 	[ObjectSystem]
-	public class MessageDispatherComponentAwakeSystem : AwakeSystem<MessageDispatherComponent>
+	public class MessageDispatcherComponentAwakeSystem : AwakeSystem<MessageDispatcherComponent>
 	{
-		public override void Awake(MessageDispatherComponent self)
+		public override void Awake(MessageDispatcherComponent t)
 		{
-			self.Awake();
+			t.Awake();
 		}
 	}
 
 	[ObjectSystem]
-	public class MessageDispatherComponentLoadSystem : LoadSystem<MessageDispatherComponent>
+	public class MessageDispatcherComponentLoadSystem : LoadSystem<MessageDispatcherComponent>
 	{
-		public override void Load(MessageDispatherComponent self)
+		public override void Load(MessageDispatcherComponent self)
 		{
 			self.Load();
 		}
@@ -25,7 +24,7 @@ namespace ETHotfix
 	/// <summary>
 	/// 消息分发组件
 	/// </summary>
-	public class MessageDispatherComponent : Component
+	public class MessageDispatcherComponent : Component
 	{
 		private readonly Dictionary<ushort, List<IMHandler>> handlers = new Dictionary<ushort, List<IMHandler>>();
 
@@ -38,10 +37,7 @@ namespace ETHotfix
 		{
 			this.handlers.Clear();
 
-			ETModel.MessageDispatherComponent messageDispatherComponent = ETModel.Game.Scene.GetComponent<ETModel.MessageDispatherComponent>();
-			ETModel.OpcodeTypeComponent opcodeTypeComponent = ETModel.Game.Scene.GetComponent<ETModel.OpcodeTypeComponent>();
-
-			List<Type> types = Game.EventSystem.GetTypes();
+			List<Type> types = Game.EventSystem.GetTypes(typeof(MessageHandlerAttribute));
 
 			foreach (Type type in types)
 			{
@@ -57,26 +53,15 @@ namespace ETHotfix
 					Log.Error($"message handle {type.Name} 需要继承 IMHandler");
 					continue;
 				}
-				
+
 				Type messageType = iMHandler.GetMessageType();
 				ushort opcode = this.Entity.GetComponent<OpcodeTypeComponent>().GetOpcode(messageType);
-				if (opcode != 0)
+				if (opcode == 0)
 				{
-					this.RegisterHandler(opcode, iMHandler);
+					Log.Error($"消息opcode为0: {messageType.Name}");
+					continue;
 				}
-
-				// 尝试注册到mono层
-				if (messageDispatherComponent != null && opcodeTypeComponent != null)
-				{
-					ushort monoOpcode = opcodeTypeComponent.GetOpcode(messageType);
-					if (monoOpcode == 0)
-					{
-						continue;
-					}
-
-					MessageProxy messageProxy = new MessageProxy(messageType, (session, o) => { iMHandler.Handle(session, o); });
-					messageDispatherComponent.RegisterHandler(monoOpcode, messageProxy);
-				}
+				this.RegisterHandler(opcode, iMHandler);
 			}
 		}
 
@@ -89,12 +74,12 @@ namespace ETHotfix
 			this.handlers[opcode].Add(handler);
 		}
 
-		public void Handle(ETModel.Session session, MessageInfo messageInfo)
+		public void Handle(Session session, MessageInfo messageInfo)
 		{
 			List<IMHandler> actions;
 			if (!this.handlers.TryGetValue(messageInfo.Opcode, out actions))
 			{
-				Log.Error($"消息 {messageInfo.Message.GetType().FullName} 没有处理");
+				Log.Error($"消息没有处理: {messageInfo.Opcode} {JsonHelper.ToJson(messageInfo.Message)}");
 				return;
 			}
