+++ Unity/Assets/Model/Module/Message/MessageDispatcherComponent.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 468d6431a15c50e42be415ce03d671c1
+guid: 1871d0e1f87ccb94781587d7a6450144
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
