--- Unity/Assets/Model/Module/Message/MessageDispatherComponent.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: f6a1591f4b7a4121a01428af98aed2c1
-timeCreated: 1503745497
