+++ Unity/Unity.Model.csproj
@@ -300,7 +300,7 @@
     <Compile Include="Assets\Model\Module\Message\IMessagePacker.cs" />
     <Compile Include="Assets\Model\Module\Message\IMHandler.cs" />
     <Compile Include="Assets\Model\Module\Message\MessageAttribute.cs" />
-    <Compile Include="Assets\Model\Module\Message\MessageDispatherComponent.cs" />
+    <Compile Include="Assets\Model\Module\Message\MessageDispatcherComponent.cs" />
     <Compile Include="Assets\Model\Module\Message\MessageHandlerAttribute.cs" />
     <Compile Include="Assets\Model\Module\Message\MessageInfo.cs" />
     <Compile Include="Assets\Model\Module\Message\MessagePool.cs" />
